requirements.txt
# BlockGuard - Python Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# ML & Data
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Database
pymongo==4.6.1

# Blockchain
web3==6.13.0
py-solc-x==2.0.2

# Visualization / Test Suite
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
